+++ libs/spark/src/TextField.ts
@@ -1,3 +1,1 @@
 export { TextField } from '@material-ui/core';
-
-export const MuiTextFieldStyleOverrides = {};